unified gradle and maven configs so that only one file needs to be changed when modifying or adding new stuff; iconFile was not present in gradle PackageTask before, added that; renamed appName to name (gradle); renamed appDescription to description (gradle);
