+++ src/main/java/io/github/fvarrui/javapackager/utils/DownloadUtils.java
@@ -1,0 +1,5 @@
+package io.github.fvarrui.javapackager.utils;
+
+public class DownloadUtils {
+
+}
